Move History providers to common providers path
