+++ client/src/components/Upload/UploadModal.test.js
@@ -7,7 +7,7 @@ import { mount } from "@vue/test-utils";
 import { getLocalVue } from "jest/helpers";
 
 import MockCurrentUser from "../providers/MockCurrentUser";
-import MockCurrentHistory from "components/History/providers/UserHistories/MockCurrentHistory";
+import MockCurrentHistory from "components/providers/History/UserHistories/MockCurrentHistory";
 
 jest.mock("app");
 jest.mock("components/providers/History/caching");
