+++ client/src/components/providers/History/CollectionContentProvider/CollectionContentProvider.js
