+++ client/src/components/providers/History/CollectionContentProvider/index.js
