+++ client/src/components/providers/History/CollectionContentProvider/watchCollection.js
