+++ client/src/components/providers/History/ContentProvider/helpers.js
