+++ client/src/components/providers/History/ContentProvider/processContentStreams.js
