+++ client/src/components/providers/History/HistoryContentProvider/contentPayload.test.js
@@ -5,9 +5,8 @@ import { ScrollPos } from "components/History/model/ScrollPos";
 import { bulkCacheContent, wipeDatabase } from "components/providers/History/caching";
 import { contentPayload } from "./contentPayload";
 import { loadContents } from "./loadContents";
-import { serverContent, testHistory, testHistoryContent } from "../../test/testHistory";
+import { serverContent, testHistory, testHistoryContent } from "components/History/test/testHistory";
 import { untilNthEmission } from "jest/helpers";
-// import { reportPayload } from "../../test/providerTestHelpers";
 
 jest.mock("app");
 jest.mock("components/providers/History/caching");
