+++ client/src/components/providers/History/HistoryContentProvider/loadContents.js
